+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 Forthcoming
 -----------
-* ...
+* [demos] log levels demo added, `#42 <https://github.com/kobuki-base/kobuki_core/issues/42>`_
 
 1.3.1 (2020-09-15)
 ------------------